+++ src/Dojang/Commands/Reflect.hs
@@ -58,9 +58,11 @@ import Dojang.Types.Context
   , FileDeltaKind (..)
   , FileEntry (..)
   , FileStat (..)
+  , IgnoredFile (..)
   , RouteMatch (..)
   , RouteState (..)
   , findMatchingRoutes
+  , getIgnoredFiles
   , getRouteState
   , makeCorrespond
   , makeCorrespondBetweenThreeFiles
@@ -85,10 +87,69 @@ reflect force allFlag _includeUnregistered _explicitSource [] = do
   -- No arguments: reflect all changed files
   ctx <- ensureContext
   pathStyle <- pathStyleFor stderr
+  codeStyle <- codeStyleFor stderr
   (allFiles, ws) <- makeCorrespond ctx
   let changedFiles = filter isChanged allFiles
   printWarnings ws
-  if null changedFiles
+
+  -- Get ignored files and warn about them
+  ignoredFiles <- getIgnoredFiles ctx
+  -- Filter to only ignored files that exist in destination
+  existingIgnored <- filterM (exists . (.destinationPath)) ignoredFiles
+  -- Create correspondences for ignored files if --force is used
+  ignoredCorrespondences <-
+    if force && not (null existingIgnored)
+      then do
+        forM existingIgnored $ \ignored -> do
+          let interPath =
+                ctx.repository.intermediatePath
+                  </> ignored.routeName
+                  </> makeRelative
+                    (normalise ignored.routeName)
+                    (normalise ignored.sourcePath)
+          makeCorrespondBetweenThreeFiles
+            interPath
+            ignored.sourcePath
+            ignored.destinationPath
+      else return []
+  let changedIgnored = filter isChanged ignoredCorrespondences
+
+  -- Print warnings for ignored files
+  unless (null existingIgnored) $ do
+    if force
+      then do
+        printStderr' Note $
+          "Including "
+            <> pack (show $ length existingIgnored)
+            <> " ignored file(s) due to "
+            <> codeStyle "--force"
+            <> ":"
+        forM_ existingIgnored $ \ignored -> do
+          printStderr $
+            "  "
+              <> pathStyle ignored.destinationPath
+              <> " (pattern: "
+              <> codeStyle (pack $ show ignored.pattern)
+              <> ")"
+      else do
+        printStderr' Warning $
+          "Skipping "
+            <> pack (show $ length existingIgnored)
+            <> " ignored file(s):"
+        forM_ existingIgnored $ \ignored -> do
+          printStderr $
+            "  "
+              <> pathStyle ignored.destinationPath
+              <> " (pattern: "
+              <> codeStyle (pack $ show ignored.pattern)
+              <> ")"
+        printStderr' Hint $
+          "Use "
+            <> codeStyle "--force"
+            <> " to include ignored files."
+
+  let allChangedFiles = changedFiles ++ changedIgnored
+  if null allChangedFiles
     then do
       printStderr "No changed files to reflect."
       return ExitSuccess
@@ -96,9 +157,9 @@ reflect force allFlag _includeUnregistered _explicitSource [] = do
       -- Display changed files
       printStderr $
         "Found "
-          <> pack (show $ length changedFiles)
+          <> pack (show $ length allChangedFiles)
           <> " changed file(s):"
-      forM_ changedFiles $ \fc -> do
+      forM_ allChangedFiles $ \fc -> do
         printStderr $ "  " <> pathStyle fc.destination.path
       -- Confirm unless --all is specified
       proceed <-
@@ -111,7 +172,7 @@ reflect force allFlag _includeUnregistered _explicitSource [] = do
               else return True -- Non-interactive: proceed
       if proceed
         then do
-          reflectCorrespondences force changedFiles
+          reflectCorrespondences force allChangedFiles
           return ExitSuccess
         else do
           printStderr "Cancelled."